fix the method for pass the test

+++ src/main/java/dev/karina/java_loops/MultiplicationTable/MultiplicationTable.java
@@ -7,7 +7,7 @@ public class MultiplicationTable {
     public static List<String> generateTable(int n){
         List<String> table = new ArrayList<>();
             for (int i = 1; i <= 10; i++){
-                table.add(n + " x " + i + " = " + (n * i));
+                table.add(n + " x " + i + " = " + ((long)n * i));
             }
             return table;
     }
